// src/app/model/team-statistics.ts
export interface TeamStatistics {
	totalOppositionHalfPasses: number;
	aerialDuelsWonPercentage: number;
	duelsWonPercentage: number;
	goalsFromOutsideTheBox: number;
	clearancesOffLine: number;
	shotsFromInsideTheBox: number;
	accurateLongBalls: number;
	fastBreaks: number;
	saves: number;
	id: number;
	tackles: number;
	totalPasses: number;
	accurateOwnHalfPassesPercentage: number;
	interceptionsAgainst: number;
	offsides: number;
	longBallsTotalAgainst: number;
	averageBallPossession: number;
	matches: number;
	accurateOppositionHalfPasses: number;
	longBallsSuccessfulAgainst: number;
	shotsFromOutsideTheBox: number;
	totalPassesAgainst: number;
	fastBreakShots: number;
	ownHalfPassesTotalAgainst: number;
	accurateOwnHalfPassesAgainst: number;
	groundDuelsWon: number;
	shots: number;
	shotsBlockedAgainst: number;
	accurateFinalThirdPassesAgainst: number;
	accurateCrosses: number;
	penaltyGoalsConceded: number;
	freeKickShots: number;
	possessionLost: number;
	corners: number;
	rightFootGoals: number;
	leftFootGoals: number;
	accurateLongBallsPercentage: number;
	bigChancesCreated: number;
	keyPassesAgainst: number;
	bigChancesAgainst: number;
	cornersAgainst: number;
	penaltiesCommited: number;
	shotsOnTargetAgainst: number;
	hitWoodwork: number;
	accuratePassesAgainst: number;
	groundDuelsWonPercentage: number;
	accuratePassesPercentage: number;
	fouls: number;
	accurateOppositionHalfPassesPercentage: number;
	bigChancesCreatedAgainst: number;
	goalsScored: number;
	bigChances: number;
	shotsFromOutsideTheBoxAgainst: number;
	accurateOppositionHalfPassesAgainst: number;
	shotsOffTargetAgainst: number;
	ownGoals: number;
	freeKickGoals: number;
	cleanSheets: number;
	shotsOnTarget: number;
	totalLongBalls: number;
	errorsLeadingToGoalAgainst: number;
	duelsWon: number;
	crossesTotalAgainst: number;
	accurateCrossesPercentage: number;
	clearances: number;
	accurateOwnHalfPasses: number;
	shotsAgainst: number;
	bigChancesMissed: number;
	shotsFromInsideTheBoxAgainst: number;
	totalGroundDuels: number;
	awardedMatches: number;
	lastManTackles: number;
	goalsConceded: number;
	tacklesAgainst: number;
	crossesSuccessfulAgainst: number;
	dribbleAttempts: number;
	bigChancesMissedAgainst: number;
	redCards: number;
	oppositionHalfPassesTotalAgainst: number;
	penaltiesTaken: number;
	yellowCardsAgainst: number;
	shotsOffTarget: number;
	goalsFromInsideTheBox: number;
	assists: number;
	dribbleAttemptsTotalAgainst: number;
	dribbleAttemptsWonAgainst: number;
	avgRating: number;
	errorsLeadingToShot: number;
	fastBreakGoals: number;
	headedGoals: number;
	totalAerialDuels: number;
	totalFinalThirdPassesAgainst: number;
	aerialDuelsWon: number;
	errorsLeadingToGoal: number;
	totalOwnHalfPasses: number;
	successfulDribbles: number;
	hitWoodworkAgainst: number;
	accuratePasses: number;
	interceptions: number;
	redCardsAgainst: number;
	penaltyGoals: number;
	totalDuels: number;
	totalCrosses: number;
	yellowRedCards: number;
	errorsLeadingToShotAgainst: number;
	clearancesAgainst: number;
	offsidesAgainst: number;
	yellowCards: number;
}